+++ apps/studio/data/oauth-secrets/client-secrets-query.ts
@@ -4,7 +4,22 @@ import { get, handleError } from 'data/fetchers'
 import { ResponseError } from 'types'
 import { clientSecretKeys } from './keys'
 
-export type ClientSecretsVariables = {
+import { components } from 'api-types'
+
+export type Secret =
+  components['schemas']['ListOAuthAppClientSecretsResponseDto']['client_secrets'][0] & {
+    client_secret?: components['schemas']['CreateOAuthAppClientSecretResponseDto']['client_secret']
+  }
+
+export type CreatedSecret = components['schemas']['CreateOAuthAppClientSecretResponseDto']
+
+export interface SecretRowProps {
+  secret: Secret
+  appId?: string
+  isNew?: boolean
+}
+
+type ClientSecretsVariables = {
   slug?: string
   appId?: string
 }
